% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{build_docker_app}
\alias{build_docker_app}
\title{build docker application}
\usage{
build_docker_app(app.name = basename(getwd()))
}
\arguments{
\item{app.name}{name of application to make (defaults to working directory)}
}
\description{
build docker application
}

