% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{find_docker_cmd}
\alias{find_docker_cmd}
\title{find the docker executable}
\usage{
find_docker_cmd()
}
\value{
character string that is location of the docker executable
}
\description{
find the docker executable
}

