% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{find_docker_image}
\alias{find_docker_image}
\title{find docker image}
\usage{
find_docker_image(repo, tag = "latest")
}
\arguments{
\item{repo}{image name}

\item{tag}{image tag (defaults to 'latest')}
}
\description{
find docker image
}

