% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_docker_app.R
\name{push_docker_app}
\alias{push_docker_app}
\title{push a docker image to AWS ECR}
\usage{
push_docker_app(app.name = basename(getwd()),
  repo = "126952269818.dkr.ecr.us-east-1.amazonaws.com")
}
\arguments{
\item{app.name}{name of application (defaults to working directory)}

\item{repo}{name of Amazon EC2 container registry}
}
\description{
push a docker image to AWS ECR
}
\details{
Amazon web services command line tools must be installed and configured for this to properly work
}

