% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_dockerize.R
\name{shiny_dockerize}
\alias{shiny_dockerize}
\title{dockerize a shiny application}
\usage{
shiny_dockerize(directory = getwd(), re.automagic = FALSE,
  app.name = basename(directory), build = TRUE, launch = TRUE,
  base = system.file("rize", "rize-Dockerfile", package = "rize"))
}
\arguments{
\item{directory}{path to directory containing the shiny app (defaults to
working directory)}

\item{re.automagic}{logical; force automagic to recreate dependencies file}

\item{app.name}{defaults to the basename of \code{directory}}

\item{build}{Should the image try to be built}

\item{launch}{Should the image be launched immediately}

\item{base}{Dockerfile template to use, use NULL to create one-line Dockerfile for customisation}
}
\description{
dockerize a shiny application
}
\details{
This is a wrapper function that uses automagic package to build,
  test, and view a dockerized shiny application. See
  \code{\link{push_docker_app}} to push the image to an Amazon EC2 container
  registry.
}
\seealso{
\code{\link{find_docker_cmd}}, \code{\link{build_docker_app}}, \code{\link{view_docker_app}}
}
