% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{view_docker_app}
\alias{view_docker_app}
\title{run and view a dockerized shiny app}
\usage{
view_docker_app(app.name = basename(getwd()))
}
\arguments{
\item{app.name}{name of application to make (defaults to working directory)}
}
\description{
run and view a dockerized shiny app
}

